Move files into a separate lib folder for better organization & Updated path aliases
