+++ tubesize/src/components/options/apiFallbackSetting.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { getFromSyncCache, setToSyncCache } from "@/cache";
+import { getFromSyncCache, setToSyncCache } from "@lib/cache";
 
 export default function ApiFallbackSetting() {
     const [apiFallback, setAPIFallback] = useState(false);
