+++ tubesize/src/components/options/optionItem.tsx
@@ -1,4 +1,4 @@
-import { setToSyncCache } from "@/cache";
+import { setToSyncCache } from "@lib/cache";
 
 export default function OptionItem({
     option,
