+++ tubesize/src/tests/utils.test.ts
@@ -1,4 +1,4 @@
-import { isYoutubePage, extractVideoTag } from "@/utils.js";
+import { isYoutubePage, extractVideoTag } from "@lib/utils.js";
 
 describe("isYoutubePage", () => {
     test("should return true for a valid YouTube URL", () => {
